Added dynamic hash chunk sizing
It was decided that the fewer comparisons at the end, the better. Thus an algorithm was added that started the hash chunk size equal to the number of hashes in the signature (hash_len, 64). The hash chunk size would then be progressively halved until at least two files matched on a single chunk.

Additional comments and small performance boosts were added by converting maps to list comprehensions.