Diff metric changed from Jaccard to abs min differences
Jaccard index was determined to be poor metric when performing final comparison on files, because larger files with more errors would get ranked above smaller files with few changes. The difference metric in diff_index simply counts the number of trigrams present in either set that isn't present in both. The minimum of this yields the two closest files.