Modified algorithm to yield more comparisons
Testing revealed that sometimes very similar files were missed when using the log-decremental hash chunk scheme. Since the goal of this program is to predictably find the two most similar files, this miss was unacceptable. Thus the decremental hash scheme was kept, but the values were changed accordingly:
Initial hash_len was set to 32 for speed, initial hash_chunk was set to 8. Files are compared as long as one hash chunk matches, regardless of the max number of hashes matched for two files.
With these changes, the probability of matching files with Jaccard indices > 0.95 approached 99%, where we needed it.
The number of comparisons at the end predictably increased. However, for most datasets we expect the list of files per hash to be small. Thus, the comparisons do not blow up to be O(n^2) unless all files are similar to each other. Only small groups of files will be compared, keeping our algorithm ~O(n).